Move quick filter drill
